// frontend/src/api/generated/models/operatorWorkLogResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface OperatorWorkLogResponse {
  id: string;
  progress_id: string;
  job_packet_id: string;
  qty_claimed: number;
  piece_rate_applied: string;
  amount_total: string;
  is_verified: boolean;
  is_paid: boolean;
}