// frontend/src/api/generated/models/workLogResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface WorkLogResponse {
  id: string;
  operator_id: string;
  qty_claimed: number;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  piece_rate_applied: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  amount_total: string;
  is_paid: boolean;
  is_verified: boolean;
  rate_adjustment_reason: string;
}